# [BUGFIX] Check Document EntryRoot respecting case sensitivity on macOS